# backend/swagger.yaml
swagger: "2.0"
info:
  description: This is the official documentation of the Pictures Manager project - EPITECH. enjoy
  version: "1.0.0"
  title: Node Rest API documentation - Picts-Manager
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: boubacar.soumare@epitech.eu
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: localhost:3002
basePath: /api/v1
schemes:
  - http
paths:
...
  /picture:
    post:
      security:
        - Bearer: []
      tags:
        - Picture Module
      summary: Picture API
      description: API for creating picture
      parameters:
        - in: body
          name: body
          description: Create Picture Payload
          required: true
          schema:
            $ref: "#/definitions/Picture"
      produces:
        - application/json
      responses:
        "200":
          description: Picture Created Successfully
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
    get:
      security:
        - Bearer: []
      tags:
        - Picture Module
      summary: Picture API
      description: API for fetching all pictures
      parameters:
        - name: skip
          in: query
          description: No.of record to be skipped for pagination
          type: number
          required: false
        - name: limit
          in: query
          description: no. of record api should return for pagination
          type: number
          required: false
      produces:
        - application/json
      responses:
        "200":
          description: Picture fetched Successfully
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
  "/picture/{id}":
    get:
      security:
        - Bearer: []
      tags:
        - Picture Module
      summary: Fetch Picture Detail
      description: Fetch Picture Detail
      parameters:
        - name: id
          in: path
          description: Picture Id
          type: string
          required: true
      produces:
        - application/json
      responses:
        "200":
          description: Picture Fetched Successfully
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
    put:
      security:
        - Bearer: []
      tags:
        - Picture Module
      summary: Update Picture
      description: Update Picture
      parameters:
        - name: id
          in: path
          description: Picture Id
          type: string
          required: true
        - in: body
          name: body
          description: Update Picture Payload
          required: true
          schema:
            $ref: "#/definitions/Picture"
      produces:
        - application/json
      responses:
        "200":
          description: Picture Updated Successfully
          schema:
            $ref: "#/definitions/ApiResponse"
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
    delete:
      security:
        - Bearer: []
      tags:
        - Picture Module
      summary: Remove Picture
      description: Remove Picture
      parameters:
        - name: id
          in: path
          description: Picture Id
          type: string
          required: true
      produces:
        - application/json
      responses:
        204:
          description: Picture Deleted Successfully
        "400":
          description: Invalid Fields
        "500":
          description: Internal Server Error
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
definitions:
...
  Picture:
    properties:
      name:
        type: string
        description: the owner id
      photo:
        type: string
        format: binary
        description: picture photo
      tags:
        type: string
        description: tags in string separated with comma (',')
...
  ApiResponse:
    type: object
    properties:
      status:
        type: integer
      message:
        type: string
      body:
        type: object